feature: export HelmetProvider along with Helmet + Plugin
In certain scenarios (such as unit testing), you might want to be able
to use a `<HelmetProvider />` without needing to set up a full fledged
fusion app instance and using the plugin.

In these cases, you currently need to add `react-helmet-async` as a
`dep/devDep` and make sure the version is not different from the one in
`fusion-plugin-react-helmet-async`.